requirements.txt
arrow==0.6.0
Flask==0.10.1
gunicorn==19.3.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-dateutil==2.4.2
requests==2.7.0
six==1.9.0
Werkzeug==0.10.4
